# lib/api-zod/src/generated/types/availableGoalsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Body Composition Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AvailableGoalsResponseAvailableGoalsItem } from "./availableGoalsResponseAvailableGoalsItem";

export interface AvailableGoalsResponse {
  availableGoals: AvailableGoalsResponseAvailableGoalsItem[];
  weightGap: number;
  /** @nullable */
  validationError?: string | null;
}