# lib/api-zod/src/generated/types/onboardingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Body Composition Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OnboardingBodyActivityLevel } from "./onboardingBodyActivityLevel";
import type { OnboardingBodyDietaryPreferencesItem } from "./onboardingBodyDietaryPreferencesItem";
import type { OnboardingBodyGender } from "./onboardingBodyGender";
import type { OnboardingBodyGoalMode } from "./onboardingBodyGoalMode";
import type { OnboardingBodyInjuryFlagsItem } from "./onboardingBodyInjuryFlagsItem";
import type { OnboardingBodyTrainingDays } from "./onboardingBodyTrainingDays";
import type { OnboardingBodyTrainingLocation } from "./onboardingBodyTrainingLocation";

export interface OnboardingBody {
  /**
   * @minimum 100
   * @maximum 250
   */
  heightCm: number;
  /**
   * @minimum 30
   * @maximum 300
   */
  weightKg: number;
  /**
   * @minimum 30
   * @maximum 300
   */
  targetWeightKg: number;
  /**
   * @minimum 16
   * @maximum 100
   */
  age: number;
  gender: OnboardingBodyGender;
  goalMode: OnboardingBodyGoalMode;
  activityLevel: OnboardingBodyActivityLevel;
  trainingDays: OnboardingBodyTrainingDays;
  trainingLocation: OnboardingBodyTrainingLocation;
  dietaryPreferences: OnboardingBodyDietaryPreferencesItem[];
  injuryFlags: OnboardingBodyInjuryFlagsItem[];
  goalOverride?: boolean;
}