# lib/api-zod/src/generated/types/profileResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Body Composition Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileResponse {
  id: number;
  heightCm: number;
  weightKg: number;
  targetWeightKg: number;
  age: number;
  gender: string;
  goalMode: string;
  activityLevel: string;
  trainingDays: number;
  trainingLocation: string;
  dietaryPreferences: string[];
  injuryFlags: string[];
  goalOverride: boolean;
}