# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Body Composition Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface SignupBody {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 8 */
  passwordConfirm: string;
  fullName?: string;
}